requirements.txt
# Core
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.1.0
langchain-ollama>=0.0.1
chromadb>=0.4.0
ollama>=0.1.0
sentence-transformers>=2.2.2

# Data
datasets>=2.14.0
beir>=2.0.0
ir_datasets>=0.5.0

# Evaluation
ragas>=0.1.0
deepeval>=0.21.0

# Utilities
pyyaml>=6.0
tqdm>=4.66.0
numpy>=1.24.0
pandas>=2.0.0
